+++ src/main/java/com/iexec/common/replicate/ReplicateStatus.java
@@ -45,10 +45,12 @@ public enum ReplicateStatus {
     REVEAL_FAILED,
     REVEALED,
     RESULT_UPLOAD_REQUESTED,
-    RESULT_UPLOAD_REQUEST_FAILED,
     RESULT_UPLOADING,
     RESULT_UPLOAD_FAILED,
     RESULT_UPLOADED,
+    CONTRIBUTE_AND_FINALIZE_ONGOING,
+    CONTRIBUTE_AND_FINALIZE_FAILED,
+    CONTRIBUTE_AND_FINALIZE_DONE,
     COMPLETING,
     COMPLETE_FAILED,
     COMPLETED,
@@ -86,6 +88,8 @@ public enum ReplicateStatus {
                 CONTRIBUTED,
                 REVEALING,
                 REVEALED,
+                CONTRIBUTE_AND_FINALIZE_ONGOING,
+                CONTRIBUTE_AND_FINALIZE_DONE,
                 COMPLETING,
                 COMPLETED));
     }
@@ -98,8 +102,8 @@ public enum ReplicateStatus {
                 COMPUTE_FAILED,
                 CONTRIBUTE_FAILED,
                 REVEAL_FAILED,
-                RESULT_UPLOAD_REQUEST_FAILED,
                 RESULT_UPLOAD_FAILED,
+                CONTRIBUTE_AND_FINALIZE_FAILED,
                 COMPLETE_FAILED,
                 FAILED);
     }
@@ -112,10 +116,10 @@ public enum ReplicateStatus {
         return Arrays.asList(
                 REVEALED,
                 RESULT_UPLOAD_REQUESTED,
-                RESULT_UPLOAD_REQUEST_FAILED,
                 RESULT_UPLOADING,
                 RESULT_UPLOAD_FAILED,
                 RESULT_UPLOADED,
+                CONTRIBUTE_AND_FINALIZE_DONE,
                 COMPLETING,
                 COMPLETE_FAILED);
     }
@@ -144,6 +148,8 @@ public enum ReplicateStatus {
                 CONTRIBUTED,
                 REVEALING,
                 REVEAL_FAILED,
+                CONTRIBUTE_AND_FINALIZE_ONGOING,
+                CONTRIBUTE_AND_FINALIZE_FAILED,
                 ABORTED);
     }
 
@@ -178,10 +184,12 @@ public enum ReplicateStatus {
                 REVEALING,
                 REVEALED,
                 RESULT_UPLOAD_REQUESTED,
-                RESULT_UPLOAD_REQUEST_FAILED,   // can complete later
                 RESULT_UPLOADING,
                 RESULT_UPLOAD_FAILED,           // can complete later
                 RESULT_UPLOADED,
+                CONTRIBUTE_AND_FINALIZE_ONGOING,
+                CONTRIBUTE_AND_FINALIZE_DONE,
+                CONTRIBUTE_AND_FINALIZE_FAILED,
                 COMPLETING,
                 RECOVERING);
     }
@@ -229,10 +237,12 @@ public enum ReplicateStatus {
                 REVEAL_FAILED,
                 REVEALED,
                 RESULT_UPLOAD_REQUESTED,
-                RESULT_UPLOAD_REQUEST_FAILED,
                 RESULT_UPLOADING,
                 RESULT_UPLOAD_FAILED,
                 RESULT_UPLOADED,
+                CONTRIBUTE_AND_FINALIZE_ONGOING,
+                CONTRIBUTE_AND_FINALIZE_FAILED,
+                CONTRIBUTE_AND_FINALIZE_DONE,
                 COMPLETING,
                 COMPLETE_FAILED);
     }
@@ -276,8 +286,8 @@ public enum ReplicateStatus {
     public static List<ReplicateStatus> getUncompletableStatuses() {
         return Arrays.asList(
                 CONTRIBUTE_FAILED,
-                REVEAL_FAILED
-                //RESULT_UPLOAD_REQUEST_FAILED, // still good if don't upload
+                REVEAL_FAILED,
+                CONTRIBUTE_AND_FINALIZE_FAILED
                 //RESULT_UPLOAD_FAILED,         //still good if don't upload
         );
     }