+++ src/main/java/com/iexec/common/sdk/order/payload/WorkerpoolOrder.java
@@ -25,6 +25,10 @@ import lombok.Value;
 
 import java.math.BigInteger;
 
+/**
+ * @deprecated Use {@link com.iexec.commons.poco.order.WorkerpoolOrder}
+ */
+@Deprecated(forRemoval = true)
 @Value
 @EqualsAndHashCode(callSuper = true)
 @ToString(callSuper = true)